requirements.txt
pyyaml
pandas
scikit-learn
#hyperopt
google-auth 
google-api-python-client
pandas-gbq
mlflow
google-cloud-storage
pyarrow
Flask~=2.0.1
prefect==2.3.1
gcsfs
gunicorn
Werkzeug~=2.0.1
requests
#prometheus_client~=0.11.0
#evidently
#pymongo
pytest
deepdiff
pylint
black
isort
pre-commit